LayoutTests/media/modern-media-controls/pip-placard/pip-placard-text-section-expected.txt

Testing that text selection is turned off for PiPPlacard.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(document.querySelector('.description')).webkitUserSelect is "none"
PASS window.getComputedStyle(document.querySelector('.description')).cursor is "default"

PASS successfullyParsed is true

TEST COMPLETE
